Massively simplified async_processor.
Now, AsyncProcessor internally uses a sender / receiver model, capped
with a BufferedUnordered over it. This means I don't have to reinvent
an async scheduler, and/or deal with the problems around them.

Additionally, process_futures() returns a AsyncProcessorHandle, which
can be used to tie together any internal information I'd like to
keep around.

As of now, this removes functionality of finding out where in the
queue each future is, but this will shortly be added back in, with
information inside the AsyncProcessorHandle struct.