[ffxiv_items] Began work item database
The Library class will be superceded with the ItemDB class.
Currently, this is intialized with a connection string in the
FFXIV_ITEM_DB environment variable. This will be changed next commit.

Currently the database isn't hooked into the server, but next commit
will attempt to hook it into the recipe endpoint.