[items] Added GitHub commit last updated checks
The CSV files that are downloaded from GitHub, from the xivapi/
ffxiv-datamining repo, were previously only re-downloaded when the
database tables were empty. Now, when starting up the server, it calls
the GitHub API for fetching commit info on a file, and determines
whether or not the file has been updated recently. If so, it
re-downloads the files and rebuilds all of the database tables with the
new data.